feat: Ported over to BlueBase
BREAKING CHANGE: Yes, will not work for BlueRain anymore. This is only for BlueBase.
