+++ README.md
@@ -1,5 +1,5 @@
 <p align="center">
-  <h1 style="text-align:center;">BlueRain Storybook Addon</h1>
+  <h1 style="text-align:center;">BlueBase Storybook Addon</h1>
 </p>
 
 ## Getting started
@@ -14,10 +14,10 @@ npm i --save-dev @blueeast/bluerain-storybook-addon
 ```javascript
 import React from 'react';
 import { configure, addDecorator } from '@storybook/react';
-import { BlueRainDecorator } from '@blueeast/bluerain-storybook-addon';
+import { BlueBaseDecorator } from '@blueeast/bluerain-storybook-addon';
 
 const BRConfigs = require('../bluerain');
-addDecorator(BlueRainDecorator(BRConfigs));
+addDecorator(BlueBaseDecorator(BRConfigs));
 
 const req = require.context('../src', true, /.stories.tsx$/);
 function loadStories() {
@@ -28,18 +28,18 @@ configure(loadStories, module);
 
 ### Usage
 
-In your story, test a BlueRain component like this:
+In your story, test a BlueBase component like this:
 
 ```typescript
-import { BlueRain, BlueRainConsumer } from '@blueeast/bluerain-os';
+import { BlueBase, BlueBaseConsumer } from '@blueeast/bluerain-os';
 import React from 'react';
 import storiesOf from '@blueeast/bluerain-storybook-addon';
 
 
 storiesOf('Some Story', module)
 	.add('story', () => (
-		<BlueRainConsumer>
-			{(BR: BlueRain) => <BR.Components.SomeComponent title="A nice component" />}
-		</BlueRainConsumer>
+		<BlueBaseConsumer>
+			{(BB: BlueBase) => <BB.Components.SomeComponent title="A nice component" />}
+		</BlueBaseConsumer>
 	));
 ```
