+++ src/index.ts
@@ -1,6 +1,5 @@
-import BlueRainDecorator from './BlueRainDecorator';
-import bluerainStory from './BlueRainStory';
+import BlueBaseDecorator from './BlueBaseDecorator';
 import storiesOf from './storiesOf';
 
-export { BlueRainDecorator, bluerainStory };
+export { BlueBaseDecorator };
 export default storiesOf;
